Conflicts:
	ganttzk/src/main/java/org/zkoss/ganttz/Planner.java
	ganttzk/src/main/resources/i18n/keys.pot
	ganttzk/src/main/resources/web/ganttz/zul/plannerLayout.zul
	ganttzk/src/main/resources/web/ganttz/zul/resourcesLoadLayout.zul
	ganttzk/src/main/resources/web/ganttz/zul/timetracker/timetrackersecondlevel.zul
	navalplanner-webapp/src/main/java/org/navalplanner/web/planner/TaskElementAdapter.java
	navalplanner-webapp/src/main/java/org/navalplanner/web/planner/company/CompanyPlanningModel.java
	navalplanner-webapp/src/main/java/org/navalplanner/web/planner/order/OrderPlanningModel.java
	navalplanner-webapp/src/main/java/org/navalplanner/web/planner/tabs/MultipleTabsPlannerController.java
	navalplanner-webapp/src/main/resources/i18n/keys.pot
	navalplanner-webapp/src/main/resources/logos
	navalplanner-webapp/src/main/webapp/calendars/_edition.zul
	navalplanner-webapp/src/main/webapp/common/configuration.zul
	navalplanner-webapp/src/main/webapp/orders/_listOrderElementAdvances.zul
	navalplanner-webapp/src/main/webapp/orders/_listOrderElementAuthorizations.zul
	navalplanner-webapp/src/main/webapp/orders/_listOrderElementHours.zul
	navalplanner-webapp/src/main/webapp/orders/_listOrderElementTaskQualityForms.zul
	navalplanner-webapp/src/main/webapp/orders/_orderFilter.zul
	navalplanner-webapp/src/main/webapp/orders/components/_listOrderElementLabels.zul
	navalplanner-webapp/src/main/webapp/orders/components/_orderElementTree.zul

+++ navalplanner-business/src/main/java/org/navalplanner/business/materials/entities/MaterialAssignment.java
@@ -91,15 +91,15 @@ public class MaterialAssignment extends BaseEntity implements Comparable {
         getMaterialInfo().setMaterial(material);
     }
 
-    public Double getUnits() {
+    public BigDecimal getUnits() {
         return getMaterialInfo().getUnits();
     }
 
-    public void setUnits(Double units) {
+    public void setUnits(BigDecimal units) {
         getMaterialInfo().setUnits(units);
     }
 
-    public void setUnitsWithoutNullCheck(Double units) {
+    public void setUnitsWithoutNullCheck(BigDecimal units) {
         this.materialInfo.setUnitsWithoutNullCheck(units);
     }
